requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
nats-py
python-jose[cryptography]
requests